+++ .yalc/@ensdomains/thorin/package.json
@@ -47,5 +47,5 @@
     "react": "^18",
     "react-dom": "^18"
   },
-  "yalcSig": "618135530556dc4c96140ba13ddce052"
+  "yalcSig": "d796d605b3db7735d0cf384941d715af"
 }
