+++ .yalc/@ensdomains/thorin/yalc.sig
@@ -1,1 +1,1 @@
-618135530556dc4c96140ba13ddce052
+d796d605b3db7735d0cf384941d715af
