+++ src/components/NFTWithPlaceholder.tsx
@@ -1,15 +1,12 @@
 import { ComponentProps } from 'react'
-import styled, { css } from 'styled-components'
 import { useEnsAvatar } from 'wagmi'
+
+import { Box, BoxProps } from '@ensdomains/thorin'
 
 import NFTTemplate from './@molecules/NFTTemplate/NFTTemplate'
 
-const StyledNftBox = styled.div(
-  ({ theme }) => css`
-    width: 100%;
-    border-radius: ${theme.radii.large};
-    overflow: hidden;
-  `,
+const StyledNftBox = (props: BoxProps) => (
+  <Box {...props} width="$full" borderRadius="$large" overflow="hidden" />
 )
 
 export const NFTWithPlaceholder = ({
@@ -23,7 +20,6 @@ export const NFTWithPlaceholder = ({
   const isCompatible = !!(name && name.split('.').length === 2 && name.endsWith('.eth'))
 
   if (!isCompatible) return null
-
   return (
     <StyledNftBox {...props}>
       <NFTTemplate name={name} backgroundImage={avatar} isNormalised />
