+++ src/components/ProfileSnippet/style.css.ts
@@ -1,0 +1,10 @@
+import { style } from '@vanilla-extract/css'
+
+import { cssVars } from '@ensdomains/thorin'
+
+export const snippet = style({
+  backgroundRepeat: 'no-repeat',
+  backgroundAttachment: 'scroll',
+  backgroundSize: `100% ${cssVars.space['28']}`,
+  backgroundPositionY: '-1px',
+})
